Supervised ff version complete 98.5% accuracy

+++ projects/forward_forward/forward_forward_sup.py
@@ -4,6 +4,7 @@ Negative examples have an incorrect labe in the input image.
 """
 # Project imports
 import sys
+
 sys.path.append("../..")
 import common.utils as common_utils
 
@@ -21,27 +22,34 @@ import matplotlib.pyplot as plt
 from tqdm import tqdm
 import h5py
 import shutil
+import argparse
+
+# Parse argments
+parser = argparse.ArgumentParser()
+parser.add_argument("--delete_output_dir", type=bool, default=False)
+args = parser.parse_args()
 
 # Comet logging
-common_utils.start_comet_ml_logging("michaelliang-dev")
-
+# common_utils.start_comet_ml_logging("michaelliang-dev")
 
 # cuda device
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
 # Load a HDF5 dataset
 
+
 def load_hdf5_dataset(file_path):
-    with h5py.File(file_path, 'r') as f:
-        input = f['input'][:]
-        target = f['target'][:]
+    with h5py.File(file_path, "r") as f:
+        input = f["input"][:]
+        target = f["target"][:]
     return input, target
 
+
 # Load the datasets
-train_input, train_target = load_hdf5_dataset('mnist_pos/train.hdf5')
-train_neg_input, train_neg_target = load_hdf5_dataset('mnist_neg/train.hdf5')
+train_input, train_target = load_hdf5_dataset("mnist_pos/train.hdf5")
+train_neg_input, train_neg_target = load_hdf5_dataset("mnist_neg/train.hdf5")
 # Load test sets
-test_input, test_target = load_hdf5_dataset('mnist_pos/test.hdf5')
+test_input, test_target = load_hdf5_dataset("mnist_pos/test.hdf5")
 
 # Move to cuda
 train_input = torch.tensor(train_input).to(device)
@@ -52,19 +60,20 @@ test_input = torch.tensor(test_input).to(device)
 test_target = torch.tensor(test_target).to(device)
 
 # Build the forward forward net
-
-# if output folder eixst
-if os.path.exists('output'):
-    shutil.rmtree('output')
-os.makedirs('output')
+if args.delete_output_dir and os.path.exists("output"):
+    shutil.rmtree("output")
+    os.makedirs("output")
+elif not os.path.exists("output"):
+    os.makedirs("output")
 
 torch.manual_seed(42)
 
 # Hyperparameters
 lr = 1e-5
 batch_size = 64
-num_epoch = 100
+num_epoch = 10
 eval_save_interval = 10000
+
 
 class ForwardForwardNet(nn.Module):
     def __init__(self):
@@ -75,22 +84,39 @@ class ForwardForwardNet(nn.Module):
         self.hidden4 = LinearBlock(in_features=2000, out_features=2000)
         self.output = nn.Linear(6000, 10)
         self.softmax = nn.Softmax(dim=1)
-        self.training = True
+        # optimizer for just the output layer
+        self.output_optimizer = torch.optim.Adam(self.output.parameters(), lr=lr)
 
     def forward(self, x_pos, x_neg):
         out1_pos, out1_neg, loss1 = self.hidden1(x_pos, x_neg)
         out2_pos, out2_neg, loss2 = self.hidden2(out1_pos, out1_neg)
         out3_pos, out3_neg, loss3 = self.hidden3(out2_pos, out2_neg)
-        out4_pos, out4_neg, loss4 = self.hidden4(out3_pos, out3_neg)
+        out4_pos, _, loss4 = self.hidden4(out3_pos, out3_neg)
 
         # # concatenate final 3 layers and put it through a classifier layer
-        # cat = torch.cat((out2, out3, out4), dim=1)
-        # logits = self.output(cat)
-        # probs = self.softmax(logits)
+        cat_pos = torch.cat((out2_pos, out3_pos, out4_pos), dim=1)
+        logits_pos = self.output(cat_pos)
+        train_target = x_pos[:, :10]
+        loss = F.cross_entropy(logits_pos, train_target)
 
-        # Avg loss
-        loss = (loss1 + loss2 + loss3 + loss4) / 4
+        self.output_optimizer.zero_grad()
+        loss.backward()
+        self.output_optimizer.step()
+
+        # average all loss values
+        loss = (loss1 + loss2 + loss3 + loss4 + loss) / 5
         return loss
+
+    def inference(self, x):
+        out1 = self.hidden1.inference(x)
+        out2 = self.hidden2.inference(out1)
+        out3 = self.hidden3.inference(out2)
+        out4 = self.hidden4.inference(out3)
+        cat = torch.cat((out2, out3, out4), dim=1)
+        logits = self.output(cat)
+        probs = self.softmax(logits)
+        return probs
+
 
 class LinearBlock(nn.Module):
     def __init__(self, in_features, out_features):
@@ -100,13 +126,15 @@ class LinearBlock(nn.Module):
         self.layernorm = nn.LayerNorm(out_features)
         self.optimizer = torch.optim.Adam(self.parameters(), lr=lr)
         self.threshold = 2.0
-        
+
     def forward(self, x_pos, x_neg):
         # Compute goodness
         g_pos = self.relu(self.linear(x_pos)).pow(2).mean(1)
         g_neg = self.relu(self.linear(x_neg)).pow(2).mean(1)
 
-        loss = torch.log(1 + torch.exp(torch.cat([-g_pos + self.threshold, g_neg - self.threshold]))).mean()
+        loss = torch.log(
+            1 + torch.exp(torch.cat([-g_pos + self.threshold, g_neg - self.threshold]))
+        ).mean()
 
         self.optimizer.zero_grad()
         loss.backward()
@@ -118,23 +146,60 @@ class LinearBlock(nn.Module):
 
         return out_pos, out_neg, loss
 
+    def inference(self, x):
+        return self.layernorm(self.relu(self.linear(x)))
+
+
 # Train loader for train_input
-train_loader = torch.utils.data.DataLoader(train_input, batch_size=batch_size, shuffle=True, drop_last=True)
-train_neg_loader = torch.utils.data.DataLoader(train_neg_input, batch_size=batch_size, shuffle=True, drop_last=True)
+train_loader = torch.utils.data.DataLoader(
+    train_input, batch_size=batch_size, shuffle=True, drop_last=True
+)
+train_neg_loader = torch.utils.data.DataLoader(
+    train_neg_input, batch_size=batch_size, shuffle=True, drop_last=True
+)
 
 
 m = ForwardForwardNet()
 m.to(device)
 
-# Train the forward forward net
-for epoch in range(num_epoch):
-    for iter, (x_pos, x_neg) in tqdm(enumerate(zip(train_loader, train_neg_loader)), total=len(train_loader)):
-        x_pos = x_pos.view(batch_size, -1).float()
-        x_neg = x_neg.view(batch_size, -1).float()
+# Load checkpoint
+m.load_state_dict(torch.load("./final_checkpoint.pt"))
 
-        loss = m(x_pos, x_neg)
+# # Train the forward forward net
+# for epoch in range(num_epoch):
+#     for iter, (x_pos, x_neg) in tqdm(
+#         enumerate(zip(train_loader, train_neg_loader)), total=len(train_loader)
+#     ):
+#         x_pos = x_pos.view(batch_size, -1).float()
+#         x_neg = x_neg.view(batch_size, -1).float()
 
-    # print and save end of every epoch
-    print(f'Epoch {epoch+1}: Loss {loss.item()}')
-    # evaluate_model()
-    torch.save(m.state_dict(), f'./output/ff_net_{epoch+1}.pt')
+#         loss = m(x_pos, x_neg)
+
+#     # print and save end of every epoch
+#     print(f"Epoch {epoch+1}: Loss {loss.item()}")
+#     # evaluate_model()
+#     torch.save(m.state_dict(), f"./output/ff_net_{epoch+1}.pt")
+
+
+# Evaluate
+m.eval()
+test_loader = torch.utils.data.DataLoader(
+    test_input, batch_size=batch_size, shuffle=False, drop_last=True
+)
+test_target_loader = torch.utils.data.DataLoader(
+    test_target, batch_size=batch_size, shuffle=False, drop_last=True
+)
+correct = 0
+total = 0
+with torch.no_grad():
+    for iter, (x, target) in tqdm(
+        enumerate(zip(test_loader, test_target_loader)), total=len(test_loader)
+    ):
+        x = x.view(batch_size, -1).float()
+        target = target.view(batch_size).long()
+        probs = m.inference(x)
+        _, pred = torch.max(probs, dim=1)
+        correct += (pred == target).sum().item()
+        total += target.shape[0]
+
+print(f"Accuracy: {correct/total}, Correct: {correct}, Total: {total}")
